+++ src/steps/resource-manager/databases/sql/constants.ts
@@ -10,6 +10,7 @@ import {
 
 export const steps = {
   DATABASES: 'rm-database-sql-databases',
+  SERVER_FIREWALL_RULES: 'rm-database-sql-server-firewall-rules',
 };
 
 export const entities = {
@@ -18,11 +19,15 @@ export const entities = {
     _class: RM_DATABASE_SERVER_ENTITY_CLASS,
     resourceName: '[RM] SQL Server',
   },
-
   DATABASE: {
     _type: 'azure_sql_database',
     _class: RM_DATABASE_ENTITY_CLASS,
     resourceName: '[RM] SQL Database',
+  },
+  FIREWALL_RULE: {
+    _type: 'azure_sql_server_firewall_rule',
+    _class: ['Firewall'],
+    resourceName: '[RM] SQL Server Firewall Rule',
   },
 };
 
@@ -30,7 +35,6 @@ export const relationships = {
   RESOURCE_GROUP_HAS_SQL_SERVER: createResourceGroupResourceRelationshipMetadata(
     entities.SERVER._type,
   ),
-
   SQL_SERVER_HAS_SQL_DATABASE: {
     _type: generateRelationshipType(
       RelationshipClass.HAS,
@@ -40,5 +44,11 @@ export const relationships = {
     sourceType: entities.SERVER._type,
     _class: RelationshipClass.HAS,
     targetType: entities.DATABASE._type,
+  },
+  SQL_SERVER_HAS_FIREWALL_RULE: {
+    _type: 'azure_sql_server_has_firewall_rule',
+    sourceType: entities.SERVER._type,
+    _class: RelationshipClass.HAS,
+    targetType: entities.FIREWALL_RULE._type,
   },
 };