+++ src/steps/resource-manager/databases/sql/index.ts
@@ -16,6 +16,7 @@ import { createDatabaseEntity, createDbServerEntity } from '../converters';
 import { SQLClient } from './client';
 import { steps, entities, relationships } from './constants';
 import {
+  createSqlServerFirewallRuleEntity,
   setAuditingStatus,
   setDatabaseEncryption,
   setServerSecurityAlerting,
@@ -105,6 +106,40 @@ export async function fetchSQLDatabases(
   });
 }
 
+export async function fetchSQLServerFirewallRules(
+  executionContext: IntegrationStepContext,
+): Promise<void> {
+  const { instance, logger, jobState } = executionContext;
+  const client = new SQLClient(instance.config, logger);
+
+  await jobState.iterateEntities(
+    { _type: entities.SERVER._type },
+    async (sqlServerEntity) => {
+      await client.iteraetServerFirewallRules(
+        {
+          id: sqlServerEntity.id as string,
+          name: sqlServerEntity.name as string,
+        },
+        async (firewallRule) => {
+          const firewallRuleEntity = await jobState.addEntity(
+            createSqlServerFirewallRuleEntity(firewallRule),
+          );
+          await jobState.addRelationship(
+            createDirectRelationship({
+              _class: RelationshipClass.HAS,
+              from: sqlServerEntity,
+              to: firewallRuleEntity,
+              properties: {
+                _type: relationships.SQL_SERVER_HAS_FIREWALL_RULE._type,
+              },
+            }),
+          );
+        },
+      );
+    },
+  );
+}
+
 export const sqlSteps: Step<
   IntegrationStepExecutionContext<IntegrationConfig>
 >[] = [
@@ -123,5 +158,13 @@ export const sqlSteps: Step<
     ],
     dependsOn: [STEP_AD_ACCOUNT, STEP_RM_RESOURCES_RESOURCE_GROUPS],
     executionHandler: fetchSQLDatabases,
+  },
+  {
+    id: steps.SERVER_FIREWALL_RULES,
+    name: 'SQL Server Firewall Rules',
+    entities: [entities.FIREWALL_RULE],
+    relationships: [relationships.SQL_SERVER_HAS_FIREWALL_RULE],
+    dependsOn: [steps.DATABASES],
+    executionHandler: fetchSQLServerFirewallRules,
   },
 ];