+++ test/helpers/getMockAccount.ts
@@ -1,0 +1,13 @@
+import { createAccountEntity } from '../../src/steps/active-directory/converters';
+import { v4 as uuid } from 'uuid';
+import { IntegrationConfig } from '../../src/types';
+
+export function getMockAccount(config: IntegrationConfig) {
+  return createAccountEntity({
+    name: 'mock-account',
+    id: uuid(),
+    integrationDefinitionId: uuid(),
+    accountId: uuid(),
+    config,
+  });
+}